backend/requirements.txt
annotated-types==0.7.0
argcomplete==3.6.2
asgiref==3.9.1
blinker==1.9.0
certifi==2025.10.5
channels==4.1.0
chardet==5.2.0
click==8.2.1
cloudinary==1.44.1
colorama==0.4.6
Django==5.2.6
django-appconf==1.1.0
django-imagekit==5.0.0
django-ninja==1.4.3
django-redis==6.0.0
dnspython==2.7.0
email_validator==2.2.0
Flask==3.1.1
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
greenlet==3.2.3
gunicorn==23.0.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
packaging==25.0
pilkit==3.0
pillow==11.2.1
pipx==1.7.1
platformdirs==4.3.8
psycopg[binary]
pydantic==2.11.9
pydantic_core==2.33.2
PyJWT==2.10.1
python-decouple==3.8
redis==5.3.1
reportlab==4.4.1
six==1.17.0
SQLAlchemy==2.0.41
sqlparse==0.5.3
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
urllib3==2.5.0
userpath==1.9.2
Werkzeug==3.1.3
WTForms==3.2.1
